pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sparse-steering"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.6.1",
    "python-dotenv>=1.0.1",
    "torch>=2.10.0",
    "transformers[torch]>=5.2.0",
    "pyyaml>=6.0",
    "tqdm>=4.0",
    "wandb>=0.25.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.setuptools.packages.find]
include = ["sparse_steer*"]
